chore: rename slide base
