+++ src/components/core/SlideBase.tsx
