+++ src/components/core/index.ts
@@ -1,2 +1,3 @@
 export * from "./SlideCore"
 export * from "./SlideFrame"
+export * from "./SlideBase"
